refactor: remove old payload classes
