--- embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/destination/SpanAttributeMapper.kt
@@ -1,8 +1,0 @@
-package io.embrace.android.embracesdk.arch.destination
-
-/**
- * Converts an object of type T to a [SpanAttributeData]
- */
-internal fun interface SpanAttributeMapper<T> {
-    fun toSpanAttributeData(obj: T): SpanAttributeData
-}
