--- embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/destination/StartSpanMapper.kt
@@ -1,8 +1,0 @@
-package io.embrace.android.embracesdk.arch.destination
-
-/**
- * Converts an object of type T to a [StartSpanData]
- */
-internal fun interface StartSpanMapper<T> {
-    fun toStartSpanData(obj: T): StartSpanData
-}
