--- embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/memory/NoOpMemoryService.kt
@@ -1,7 +1,0 @@
-package io.embrace.android.embracesdk.capture.memory
-
-internal class NoOpMemoryService : MemoryService {
-
-    override fun onMemoryWarning() {
-    }
-}
