+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/FakeMemoryService.kt
@@ -1,9 +1,8 @@
 package io.embrace.android.embracesdk.fakes
 
 import io.embrace.android.embracesdk.capture.memory.MemoryService
-import io.embrace.android.embracesdk.payload.MemoryWarning
 
-internal class FakeMemoryService : FakeDataCaptureService<MemoryWarning>(), MemoryService {
+internal class FakeMemoryService : MemoryService {
 
     var callCount = 0
