--- embrace-android-sdk/src/test/resources/memory_warning_expected.json
@@ -1,3 +1,0 @@
-{
-  "ts": 16098234098234
-}
